linux(04)iptables.yml: handle chroot environments
In chroot environments using iptabels and iptabels-save
command will cause unreliable results. The iptables rules
are not written to disk, which is required for image
building in chroot. Using iptables-save to store the set
rules to the config file includes also all rules of the
build host of the chroot environment.

Adapted code to handle chroot and write iptables settings
directly to the persistent file.

Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>